requirements.txt
rdkit-pypi==2022.9.5
torch==2.0.1
transformers==4.30.0
flask==2.2.5
werkzeug==2.2.3
neo4j==5.5.0
scikit-learn==1.2.2
chembl-webresource-client==0.10.8
pysmiles==1.0.1
python-dotenv
PyPDF2
numpy<2
